+++ src/Sulu/Bundle/WebsiteBundle/Twig/Seo/SeoTwigExtension.php
@@ -12,6 +12,7 @@ namespace Sulu\Bundle\WebsiteBundle\Twig\Seo;
 
 use Sulu\Bundle\WebsiteBundle\Twig\Content\ContentPathInterface;
 use Sulu\Component\Webspace\Analyzer\RequestAnalyzerInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
 
 /**
  * This twig extension provides support for the SEO functionality provided by Sulu.
@@ -28,9 +29,18 @@ class SeoTwigExtension extends \Twig_Extension
      */
     private $contentPath;
 
-    public function __construct(RequestAnalyzerInterface $requestAnalyzer, ContentPathInterface $contentPath)
-    {
+    /**
+     * @var RequestStack
+     */
+    private $requestStack;
+
+    public function __construct(
+        RequestAnalyzerInterface $requestAnalyzer,
+        ContentPathInterface $contentPath,
+        RequestStack $requestStack
+    ) {
         $this->requestAnalyzer = $requestAnalyzer;
+        $this->requestStack = $requestStack;
         // FIXME Should not use another twig extension here, that is not the intended use case of twig extensions
         $this->contentPath = $contentPath;
     }
@@ -73,10 +83,18 @@ class SeoTwigExtension extends \Twig_Extension
             $webspaceKey = $webspace->getKey();
         }
 
+        $request = $this->requestStack->getCurrentRequest();
+
         $html .= $this->renderTitle($seoExtension, $content);
         $html .= $this->renderMetaTags($seoExtension);
-        $html .= $this->renderAlternateLinks($urls, $webspaceKey);
-        $html .= $this->renderCanonicalTag($seoExtension, $urls, $shadowBaseLocale, $webspaceKey);
+        $html .= $this->renderAlternateLinks($urls, $webspaceKey, $request->getScheme());
+        $html .= $this->renderCanonicalTag(
+            $seoExtension,
+            $urls,
+            $shadowBaseLocale,
+            $webspaceKey,
+            $request->getScheme()
+        );
 
         return $html;
     }
@@ -161,12 +179,14 @@ class SeoTwigExtension extends \Twig_Extension
      *
      * @param string[] $urls All the localized URLs for the current page
      * @param string $webspaceKey The key of the current webspace
+     * @param string $scheme scheme of request (http or https)
      *
      * @return string The rendered HTML tags
      */
-    private function renderAlternateLinks(array $urls, $webspaceKey)
+    private function renderAlternateLinks(array $urls, $webspaceKey, $scheme)
     {
         $html = '';
+        $concreteTranslations = 0;
 
         $defaultLocale = null;
         $portal = $this->requestAnalyzer->getPortal();
@@ -179,11 +199,17 @@ class SeoTwigExtension extends \Twig_Extension
             // the only exception is the homepage where the requested resource-locator is false
             if ($url !== '/' || $this->requestAnalyzer->getResourceLocator() === false) {
                 if ($defaultLocale === $locale) {
-                    $html .= $this->renderAlternateLink($url, $webspaceKey, $locale, true);
+                    $html .= $this->renderAlternateLink($url, $webspaceKey, $locale, true, $scheme);
                 }
 
-                $html .= $this->renderAlternateLink($url, $webspaceKey, $locale);
+                $html .= $this->renderAlternateLink($url, $webspaceKey, $locale, false, $scheme);
+                $concreteTranslations++;
             }
+        }
+
+        // if only one translation is available dismiss hreflang tags
+        if ($concreteTranslations <= 1) {
+            return '';
         }
 
         return $html;
@@ -196,14 +222,15 @@ class SeoTwigExtension extends \Twig_Extension
      * @param string $webspaceKey The key of the current webspace
      * @param string $locale The locale for which the tag should be rendered
      * @param bool $default If true the tag will be rendered as default locale
+     * @param string $scheme scheme of request (http or https)
      *
      * @return string The rendered alternate link tag
      */
-    private function renderAlternateLink($url, $webspaceKey, $locale, $default = false)
+    private function renderAlternateLink($url, $webspaceKey, $locale, $default = false, $scheme = 'http')
     {
         return sprintf(
             '<link rel="alternate" href="%s" hreflang="%s"/>' . PHP_EOL,
-            rtrim($this->contentPath->getContentPath($url, $webspaceKey, $locale), '/'),
+            rtrim($this->contentPath->getContentPath($url, $webspaceKey, $locale, null, $scheme), '/'),
             $default ? 'x-default' : str_replace('_', '-', $locale)
         );
     }
@@ -217,10 +244,11 @@ class SeoTwigExtension extends \Twig_Extension
      * @param string[] $urls All the localized URLs for the current page
      * @param string $shadowBaseLocale The displayed language, in case the current page is a shadow
      * @param string $webspaceKey The key of the current webspace
+     * @param string $scheme scheme of request (http or https)
      *
      * @return string The rendered canonical link tag
      */
-    private function renderCanonicalTag(array $seoExtension, array $urls, $shadowBaseLocale, $webspaceKey)
+    private function renderCanonicalTag(array $seoExtension, array $urls, $shadowBaseLocale, $webspaceKey, $scheme)
     {
         $canonicalTagHtml = '<link rel="canonical" href="%s"/>' . PHP_EOL;
 
@@ -231,7 +259,13 @@ class SeoTwigExtension extends \Twig_Extension
         if ($shadowBaseLocale && isset($urls[$shadowBaseLocale])) {
             return sprintf(
                 $canonicalTagHtml,
-                $this->contentPath->getContentPath($urls[$shadowBaseLocale], $webspaceKey, $shadowBaseLocale)
+                $this->contentPath->getContentPath(
+                    $urls[$shadowBaseLocale],
+                    $webspaceKey,
+                    $shadowBaseLocale,
+                    null,
+                    $scheme
+                )
             );
         }